+++ contracts/rewards/RewardsProtocol.sol
@@ -5,9 +5,9 @@ pragma solidity ^0.4.11;
 /// @notice This is to be considered as a protocol on how to access the underlying Rewards Contract
 contract RewardsProtocol {
 
-  // CONSTANT METHODS
+    // CONSTANT METHODS
 
-  function calculateManagementReward(uint timeDifference, uint gav) constant returns (uint) {}
-  function calculatePerformanceReward(uint sharePriceDifference, uint totalSupply) constant returns (uint) {}
+    function calculateManagementReward(uint timeDifference, uint gav) constant returns (uint) {}
+    function calculatePerformanceReward(uint sharePriceDifference, uint totalSupply) constant returns (uint) {}
 
 }