+++ migrations/3_deploy_participation.js
@@ -1,9 +1,0 @@
-const Subscribe = artifacts.require('./Subscribe.sol');
-const Redeem = artifacts.require('./Redeem.sol');
-
-module.exports = (deployer) => {
-  deployer.deploy([
-    Subscribe,
-    Redeem,
-  ]);
-};