+++ migrations/6_deploy_gov.js
@@ -1,7 +1,7 @@
 const Version = artifacts.require('./Version.sol');
 const Governance = artifacts.require('./Governance.sol');
 
-module.exports = (deployer) => {
+module.exports = async (deployer) => {
   // Deploy meta strucutre
-  deployer.deploy(Governance).then(() => deployer.deploy(Version, Governance.address));
+  await deployer.deploy(Governance).then(() => deployer.deploy(Version, Governance.address));
 };