+++ migrations/config/exchange_info.js
@@ -1,0 +1,11 @@
+module.exports = {
+  kovan: [
+    {
+      address: '',
+      name: 'OasisDex',
+      data: '',
+      url: 'https://oasisdex.com',
+      githubHint: '',
+    },
+  ],
+};